Update README.md
A GUI to pentest wifi Network, based on Aircrack-ng tools. This project is in initial phase and I am expecting it to have lots of bugs. But this kind of GUI project is kind of requirement as AirCrack set of tools are not yet fully utlized.

I am expecting support from the InfoSec Community to help me to grow this project.

+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++ To run this tools, Type the following command: sh runme.sh